DA#137: Modified: Separate Indexes and Collection Directory to provide different functionality
